+++ src/formats/eszip.ts
@@ -5,6 +5,7 @@ import { DenoBridge } from '../bridge.js'
 import type { Bundle } from '../bundle.js'
 import { wrapBundleError } from '../bundle_error.js'
 import { EdgeFunction } from '../edge_function.js'
+import { ImportMap } from '../import_map.js'
 import { getFileHash } from '../utils/sha256.js'
 
 interface BundleESZIPOptions {
@@ -14,14 +15,7 @@ interface BundleESZIPOptions {
   deno: DenoBridge
   distDirectory: string
   functions: EdgeFunction[]
-}
-
-const bundle = async (options: BundleESZIPOptions) => {
-  try {
-    return await bundleESZIP(options)
-  } catch (error: unknown) {
-    throw wrapBundleError(error, { format: 'eszip' })
-  }
+  importMap: ImportMap
 }
 
 const bundleESZIP = async ({
@@ -31,6 +25,7 @@ const bundleESZIP = async ({
   deno,
   distDirectory,
   functions,
+  importMap,
 }: BundleESZIPOptions): Promise<Bundle> => {
   const extension = '.eszip'
   const destPath = join(distDirectory, `${buildID}${extension}`)
@@ -39,6 +34,7 @@ const bundleESZIP = async ({
     basePath,
     destPath,
     functions,
+    imports: importMap.imports,
   }
   const flags = ['--allow-all']
 
@@ -46,7 +42,11 @@ const bundleESZIP = async ({
     flags.push('--quiet')
   }
 
-  await deno.run(['run', ...flags, bundler, JSON.stringify(payload)], { pipeOutput: true })
+  try {
+    await deno.run(['run', ...flags, bundler, JSON.stringify(payload)], { pipeOutput: true })
+  } catch (error: unknown) {
+    throw wrapBundleError(error, { format: 'eszip' })
+  }
 
   const hash = await getFileHash(destPath)
 
@@ -61,4 +61,4 @@ const getESZIPBundler = () => {
   return bundlerPath
 }
 
-export { bundle }
+export { bundleESZIP as bundle }